+++ rat-salsa/examples/async1.rs
@@ -8,7 +8,7 @@ use dirs::cache_dir;
 use rat_salsa::poll::PollTokio;
 use rat_salsa::poll::{PollCrossterm, PollRendered, PollTasks, PollTimers};
 use rat_salsa::{run_tui, RunConfig};
-use rat_theme2::schemes::IMPERIAL;
+use rat_theme2::palettes::IMPERIAL;
 use rat_theme2::DarkTheme;
 use std::fs;
 use std::fs::create_dir_all;