+++ rat-salsa/examples/minimal.rs
@@ -5,7 +5,7 @@ use crate::scenery::{Scenery, SceneryState};
 use anyhow::Error;
 use rat_salsa::poll::{PollCrossterm, PollRendered, PollTasks, PollTimers};
 use rat_salsa::{run_tui, RunConfig};
-use rat_theme2::schemes::IMPERIAL;
+use rat_theme2::palettes::IMPERIAL;
 use rat_theme2::DarkTheme;
 use std::fs;