+++ rat-theme2/src/lib.rs
@@ -1,43 +1,17 @@
-use crate::schemes::{
-    BASE16, BASE16_RELAXED, IMPERIAL, MONEKAI, MONOCHROME, OCEAN, OXOCARBON, RADIUM, TUNDRA,
-    VSCODE_DARK,
+use crate::palettes::{
+    BASE16, BASE16_RELAXED, BLACKWHITE, IMPERIAL, MONEKAI, MONOCHROME, OCEAN, OXOCARBON, RADIUM,
+    TUNDRA, VSCODE_DARK,
 };
 
-mod base16;
-mod base16r;
-mod blackwhite;
 mod dark_theme;
-mod imperial;
-mod monekai;
-mod monochrome;
-mod ocean;
-mod oxocarbon;
-mod radium;
-mod scheme;
-mod tundra;
-mod vscode_dark;
+mod palette;
+pub mod palettes;
 
-use crate::blackwhite::BLACKWHITE;
 pub use dark_theme::*;
-pub use scheme::*;
+pub use palette::*;
 
-/// Color schemes
-pub mod schemes {
-    pub use crate::base16::BASE16;
-    pub use crate::base16r::BASE16_RELAXED;
-    pub use crate::blackwhite::BLACKWHITE;
-    pub use crate::imperial::IMPERIAL;
-    pub use crate::monekai::MONEKAI;
-    pub use crate::monochrome::MONOCHROME;
-    pub use crate::ocean::OCEAN;
-    pub use crate::oxocarbon::OXOCARBON;
-    pub use crate::radium::RADIUM;
-    pub use crate::tundra::TUNDRA;
-    pub use crate::vscode_dark::VSCODE_DARK;
-}
-
-/// All currently existing color schemes.
-pub fn color_schemes() -> Vec<(String, Scheme)> {
+/// All currently existing color palettes.
+pub fn color_palettes() -> Vec<(String, Palette)> {
     vec![
         ("Black&White".to_string(), BLACKWHITE),
         ("Imperial".to_string(), IMPERIAL),
@@ -53,7 +27,7 @@ pub fn color_schemes() -> Vec<(String, Scheme)> {
     ]
 }
 
-/// A list of DarkTheme for all color schemes.
+/// A list of DarkTheme for all color palettes.
 pub fn dark_themes() -> Vec<DarkTheme> {
     vec![
         DarkTheme::new("Black&White".to_string(), BLACKWHITE),